third_party/blink/web_tests/external/wpt/css/css-animations/parsing/animation-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Default animation value
  assert_equals: expected "0s ease 0s 1 normal none running none" but got "none 0s ease 0s 1 normal none running"
Harness: the test ran to completion.
